Patches the code for compatibility with the latest file-loader
Specifically, it switches from "asset-require-hook" to "loader-utils"
for the asset URL generation, and ensures the name config like
"/path/[contenthash].[ext]" generates the same URL as Webpack's
"file-loader", which was not the case with the upstream library with
Wepback@5 and latest "file-loader". This update does not break any
existing tests in the plugin, but probably may break some edge cases.
However, it should be pretty simple to fix any particular issues.